get sub as it was

+++ backend/users/views.py
@@ -18,22 +18,21 @@ class LimitPageNumberPagination(pagination.PageNumberPagination):
     page_size_query_param = 'limit'
 
 
-class UserViewSet(UserViewSet):
+class UserViewSet(UserViewSet): 
     pagination_class = LimitPageNumberPagination
 
     @action(detail=True, methods=('post',),
             permission_classes=(IsAuthenticated,))
     def subscribe(self, request, id=None):
-        user = request.user
+        user = request.user 
         author = get_object_or_404(User, id=id)
 
-        serializer = SubscribeSerializer(
-            data={'author': author},
-            context={'request': request})
-        serializer.is_valid(raise_exception=True)
-
-        if user.subscriptions.filter(author=author).exists():
+        if user == author:
             return Response({
+                'errors': 'Вы не можете подписываться на самого себя.'
+            }, status=status.HTTP_400_BAD_REQUEST)
+        if Subscription.objects.filter(user=user, author=author).exists():
+            return Response({ 
                 'errors': 'Вы уже подписаны на данного пользователя.'
             }, status=status.HTTP_400_BAD_REQUEST)
 
@@ -45,30 +44,28 @@ class UserViewSet(UserViewSet):
 
     @subscribe.mapping.delete
     def delete_subscribe(self, request, id=None):
-        user = request.user
+        user = request.user 
         author = get_object_or_404(User, id=id)
-
-        serializer = SubscribeSerializer(
-            data={'author': author},
-            context={'request': request})
-        serializer.is_valid(raise_exception=True)
-
-        follow = user.subscribers.filter(author=author)
-        if follow.exists():
-            follow.delete()
+        if user == author: 
+            return Response({ 
+                'errors': 'Вы не можете отписываться от самого себя.'
+            }, status=status.HTTP_400_BAD_REQUEST)
+        follow = Subscription.objects.filter(user=user, author=author)
+        if follow.exists(): 
+            follow.delete() 
             return Response(status=status.HTTP_204_NO_CONTENT)
-        return Response({
+        return Response({ 
             'errors': 'Вы не подписаны на этого автора.'
         }, status=status.HTTP_400_BAD_REQUEST)
 
-    @action(detail=False, permission_classes=(IsAuthenticated,))
+    @action(detail=False, permission_classes=(IsAuthenticated,)) 
     def subscriptions(self, request):
         user = request.user
         queryset = Subscription.objects.filter(user=user)
         pages = self.paginate_queryset(queryset)
         serializer = SubscribeSerializer(
             pages,
-            many=True,
+            many=True, 
             context={'request': request}
         )
         return self.get_paginated_response(serializer.data)
